Bug 879903 - Use a weighted average with linear fall-off [r=catlee]
This patch avoids "spooky action at a distance" where an abrupt change in the
moving average is caused by a point far from the one being analyzed.  The
weighted average gives heavier weights to points near the point being
analyzed, with a gradual falloff for more distant points.  This smooths out
the effect of any changes not at the "interesting" edge of the window.