change the build system type detection to not rely on the ref data signature
The logic to detect the build system type was based on the reference
data signature on the job table to match at least one signature on the
reference_data_signatures table. This is not always true as reported in
https://bugzilla.mozilla.org/show_bug.cgi?id=1033107.
The new logic is inspecting the reference_data_signature table to
retrieve a dictionary of repo => build system type that is cached.