Bug 1060339 - Parse logs for error lines even if the job was successful
In an ideal world, we would never have a successful job that contained
errors which our log parser regex matched against. Unfortunately this is
not the case in practice, since:

1) There is sometimes spammy log output that isn't intended to make the
   job fail, but which false-positive matches against our log regex.
2) Sometimes the test/harness/buildbot/... is broken and the log error
   is real, and it should have caused the run to fail.

In the case of #1, hiding these errors in successful runs means that
when someone looks at the error summary for a failed run, they don't
realise that a proportion of the failures they can see are actually
present in _every_ job, not just the failed ones.

In the case of #2, hiding the errors means we don't realise <foo> is
broken and needs fixing.

As such, we now parse for error lines in all jobs, including successful.