Bug 1086896 - Use the log step finish line for determining result
For buildbot jobs there are headers/separators at the start and end of
each step within the job log, which contain various stats about the
step. Since buildbot buffers the output, it's able to add the result for
the current step to not just the end separator, but the start one too.

However, for non-buildbot submitters of results, they may not buffer the
output, and so cannot output the step result in the start line, only the
end one. As such, this commit switches the log parser to using the
latter instead of the former to determine the step result.