Bug 1125464 - Fix key name for specifying master vs read-only DB host
The 'host' property is unrecognised (and ignored) by Datasource, causing
it to silently fall back to the default of 'master_host'. The property
we needed to have set is in fact called 'host_type'.

As a result, we never use the read-only host!

An issue has been filed against Datasource for making this less silent:
https://github.com/jeads/datasource/issues/20