Bug 1126356 - Add support for |manage.py ingest_push <project> tip|
The most recent push from a repository is given the tag 'tip' in
Mercurial. To save having to look up a specific revision to ingest when
testing, 'tip' can now be passed to ingest_push, and the revision that
'tip' currently maps to, will be used during push and jobs ingestion.
Previously, jobs ingestion would have used the non-existent SHA of 'tip'
when trying to associate jobs with pushes, resulting in no jobs being
imported from builds-{pending,running,4hr}.