Bug 1131560 - Use simplejson inside production code
It appears as if simplejson is a fair bit faster than the standard json module,
which can make a difference in some cases.

For example, with the performance data, we can wind up processing a fair bit
of json. Each time isn't huge (0.07s on my laptop for the biggest series
over 90 days), but in aggregate (e.g. 51 series in tp5 * 5 time series)
that could add up. simplejson is nearly 10 times (0.009s for the same
series) and we already have it installed, so let's just use it. ujson is
slightly faster still (0.006s) but it's not installed.