Bug 1078415 - Log parser: Return early in parse_line() to simplify flow
If we're waiting for a step start line, we can just return early,
regardless of whether we found one or not, which means the later
check for |self.state == self.ST_STARTED| is now redundant. The previous
flow and comments were making the self.sub_parser.parse_line() call
look like it only occurred on the end step, when in fact it occurs only
when it's not the start _or_ end step.