Bug 1133455 - Fix the exception when falling back to requesttimes
In bug 1093743, the handling of request times was simplified (since we
now use a single value from request_times rather than all of them).
Unfortunately in the process, this broke the handling of cases where
request_times does not exist, and we have to fall back to requesttime.
The latter is an int, whereas the former a dict, so in addition to
fixing the exception, we now no longer try to store both in the same
variable, to avoid future confusion.