Bug 1134804 - Only run cycle-data against active repositories
init_datasources only creates repositories that have an active_status of
'active'. This means that if an environment was set up after a
repository was marked as in-active, the databases for it will not exist
and so cause exceptions when we attempt to run data cycling against
them. Bug 1035294 will take care of removing the unused databases, so we
should just exclude them from the standard data-cycling.