Bug 1144916 - Vagrant: Don't use out of date pip/virtualenv packages
We use an Ubuntu Precise image for the Vagrant project, whose package
repository contains extremely out of date versions of setuptools, pip
and virtualenv. So we stop using the package repository, and instead use
the official PyPA bootstrap script, which installs the latest versions
of setuptools and pip, and then we follow that with a pip install of the
latest virtualenv. We need curl to fetch the pip bootstrap script,
but it's handy for development anyway.

References to distribute are removed, since it's an older fork of
setuptools and long since deprecated.

Shiny!