Bug 1151629 - Don't limit the number of bugs retrieved by fetch_bugs
Previously we limited the number of pages to 30, which with a page size
of 500, meant a max of 15,000 intermittent-failure bugs retrieved from
Bugzilla, with no exceptions or log output to indicate this had
occurred.

We now have more than 15,000 intermittent failure bugs, so the limit is
being removed, since we're both confident that the search terms are
correct, and any other infinite loop would be caught by the existing
600s timeout.

The task currently takes ~300s to run, so there is still plenty of
headroom. Plus a timeout exception would be immediately visible in New
Relic and so much less of a pain to debug.