Bug 1146184 - Clean up the directory variables in runner scripts
The existing name "curr_dir" isn't overly clear - does it mean the
current working directory or the directory in which this script exists?
It should also be uppercase. I think the behaviour is clearer when
reworked like this :-)

There's a lot of other cleanup that could be done (eg reducing
duplication), but saving that for bug 1153971.