Bug 1155647 - Don't generate log parser exceptions if the log URL 404s
If attempting to fetch the log from the provided URL results in a 404,
mark the parse_status as failed but do not generate an exception. Whilst
missing logs are not ideal, they are not an app error, nor something is
likely to change over time (vs say an HTTP 500 from ftp.mozilla.org).