Bug 1145966 - Correct error message for invalid 'repo' in URL
If the 'repo' value specified in the URL does not exist, an appropriate
error message should be shown. This was working fine before if only
'repo' was set in the URL, but not if the 'revision' parameter was also
set. Now, we correctly show the "unknown repo" error for the latter case
as well.