Bug 1158893 - Delete any existing Cython build output as part of deploy
The Cython build process creates .c and .so files in the log_parser
directory, which only get re-created if the original source changes.
We recently updated the version of Cython, however this doesn't trigger
re-building them. To avoid surprises, we should purge them like we do
for the .pyc files.

In addition, there is no need to delete .pyc files from the UI source
directory, since that does not contain any Python files.