Bug 1126943 - Delete jobs from the objectstore table once ingested
Rather than retaining a job in the objectstore and giving it a
'processed_state' of 'completed', we now delete the job from the
objectstore. This reduces data usage, as well as improves the
performance of the table, since it will now only contain at most a
thousand or so rows, rather than millions.

Many tests had to be updated, since they assumed that a job would
persist in the objectstore after being successfully ingested. Rather
than counting the number of rows with a 'processed_state' of 'completed'
we can now just check there are no rows left at all in the objectstore,
which covers both the 'pending' and error cases.