Bug 1126943 - Expire data from the objectstore independently of jobs DB
Items in the objectstore are currently expired by finding the list of
result sets matching the date range, then looking up the jobs for those
result sets, and finally searching for matching job guids in the
datastore table. This is not only bad for performance of objectstore
deletes (since we end up with lists of thousands of guids), but also
means we cannot set a different cycle interval for the objectstore.

The new approach is much simpler: we only query the objectstore, and use
loaded_timestamp to determine which rows to cycle. The objectstore does
not have any foreign keys, so this isn't a problem. The only constraint
is that we must keep the complete jobs long enough for the job to stop
appearing in builds-4hr, to prevent us from continually re-adding it to
the objectstore. For now, we also only cycle jobs with a processed_state
of 'complete', so entries with errors (or that are stuck in the
'loading' state due to bug 1125476) are not lost (this matches the prior
behaviour, since the list of job_guids would only include successfully
ingested jobs).

For now the objectstore cycle interval has been set to the same default
interval as the jobs tables, but this will be reduced once manual cycle
data runs are run on stage/prod.