Bug 1169481 - remove dead resultStatusFilters/List code
We used to have a feature where you could toggle jobs on and off at each resultset
by clicking the result type.  We removed that feature a long time ago.
But still... it's legacy remains.  Remove that legacy.

There are still places where resultStatusFilters are mentioned.  But in all the remaining cases,
those pertain to the global filters, not these per-resultset ones.