Bug 1176482 - WhiteNoise: Serve index.html for bare '/' URIs too
Customise the DjangoWhiteNoise class, so that we add additional dummy
'/' entries to the file mapping, for each index.html static file found.
We also append the index filename to the URL in find_file(), since in
debug mode files are served directly from the filesystem instead of
using the list in `self.files`.

This means we can remove the site root redirect, allowing the Treeherder
homepage to be served from:
https://treeherder-heroku.herokuapp.com/
...rather than:
https://treeherder-heroku.herokuapp.com/index.html

The original add_files() method:
https://github.com/evansd/whitenoise/blob/7de97f915479c46834ba7dcaadbd58e3cdc24b40/whitenoise/base.py#L165-L178

The original find_file() method:
https://github.com/evansd/whitenoise/blob/7de97f915479c46834ba7dcaadbd58e3cdc24b40/whitenoise/django.py#L85-L90