Bug 1181572 - Stop calculating/storing pending_eta
Since bug 1096605, the UI no longer uses pending_eta, since it's too
hard to predict, and thus unreliable. As such, we can remove it from the
API response & stop calculating/storing it during ingestion.

Once this is deployed, we can delete all rows in the job_eta tables that
have a 'state' of 'pending'. After that, we can remove the 'state' field
and perform some further simplification of ETA handling (but that cannot
be done now, or we'll inadvertently include pending rows in the running
ETA stats).