Bug 1182299 - Support custom log name param in logslice
This adds the ability to specify a custom log name and have the log
viewer use the ``logname`` param of the ``text_log_summary`` to get the
right log.

This also improves the error message returned by the /logslice/ API if a
log name is used that is not found.