Bug 1178240 - Remove buildapi *JobsProcess extracted_content conditional
Checking that running JsonExtractorMixin.extract() on
builds-{pending,running,4hr} returns a Truthy result isn't something we
should do. The only case it catches is if the datasources contain only
the string "{}" (since inability to download or json.loads() would
instead result in an exception). This case is unlikely (since the file
is more likely to contain an empty set of top-level keys if there were
no jobs) and is also something we shouldn't silently ignore.

An exception would alert us more quickly that the datasources are empty.