Bug 1187267 - Handle invalid builds-4hr blobber_files json gracefully
The blobber_files property on jobs in builds-4hr is a json blob,
containing a number of name-url pairs. In some cases this blob is
truncated, so fails to decode. This change makes us handle these
failures more gracefully (by treating it as though the blobber_files
property was not set for that job), rather than causing the entire
builds-4hr task to fail.

Bug 1187284 is filed to try and make mozharness never output truncated
json in the first place.