Bug 1191080 - Initialise the New Relic agent correctly
The agent is supposed to be initialised before any other imports. In
addition, we're supposed to use WSGIApplicationWrapper() if we're not
using it as a decorator.

This unfortunately doesn't fix bug 1191080 itself, or answer the
question of whether we should instead be using the wrapper scripts
(which incidentally don't fix the bug either), but if we're going to
initialise the agent manually, then at least let's do it the way the
docs recommend.

See:
https://docs.newrelic.com/docs/agents/python-agent/installation-configuration/python-agent-integration#wsgi-application
https://docs.newrelic.com/docs/agents/python-agent/customization-extension/python-instrumentation-api#WSGIApplicationWrapper