Bug 1188444 - Save the last step in a log even if no finish marker seen
Log steps are updated/marked as finished whenever a finish marker is
seen (or the next start marker is encountered). In the case of logs
where the finish marker is missing for the very last step, we previously
did not update the result/finish time/duration/all_errors for that step,
causing any errors within it to be missing from the error summary. This
occurs frequently for Taskcluster infrastructure failures.

Each parser can now specify tasks that should be run at the end of
parsing, by overriding ParserBase.finish_parse(). The StepParser uses
this to save/update the final step if the final finish marker was
missing from the log.