First cut at autostar / intermittent orange detection.
This adds an autostar command and a detect_intermittents command.  The
former is designed to take an incoming job with an error summary and
look for existing results marked as intermittent that are a close
match for the new result. At present only two matchers are
implemented; one that requires an exact match in terms of tests, and
one that requires an exact match in terms of logs. Matching is also
constrained to be based on single lines; it is anticipated that future
iterations may add support for matching on groups of lines.

The detect_intermittents command is designed to take a group of jobs
running on the same push and with the same build job (i.e. same
testsuite, same chunk, etc.) and look for new intermittents to add to
the database. This currently only looks for test failures where there
is at least one green job and one non-green job.

There is currently no UI for seeing matches or for adding new
prototypical intermittents as match candidates. There is also no
integration with bugzilla; future development should add association
of frequent intermittents with bugs.