Bug 1147676 - Exclude local.conf.js from grunt build to avoid 404s
The parameter after "build:" is the block type, which is normally 'js',
'css' or a custom type. If we use a non-existent custom type (such as
"dontbuild"), grunt-usemin will just exclude the block, which is just
what we need to exclude this development-only file from the generated
assets used on stage/production.

See:
https://github.com/yeoman/grunt-usemin#blocks