Bug 1181572 - Stop retrieving rows of type 'pending' from job_eta table
Now that we no longer store pending_eta for each job in the jobs table,
there is no point fetching pending job related rows, when
get_last_eta_by_signatures is used as part of jobs ingestion. There is
much more simplification that could be performed (ie not returning as
nested an object in get_job_eta_times() now that we only return one
state "running") - but job_eta will be rewritten entirely soon, so it's
not worth it.