Bug 1197827 - Treeherder unclassified failure count is off
The issue with this bug was grouping, as it turns out.  ``resultset_store`` was grouping jobs by their ``name`` and ``symbol``.  However, we had the same ``symbol`` with two different names (one from Task Cluster, one from Buildbot).  In this case, it was the ``?`` group which shows as not within any group.

Since the name from each source was different, we had 2 groups both with the symbol of ``?``.  We would render the first one, then see another group named ``?`` and empty the first one, and add the new jobs.

What I've done is to fix both parts of that.

1. We group in ``resultset_store`` by *only* the group symbol to prevent conflicts there.
2. But if we DO, by magic, get two groups with the same symbol, then ``clonejobs.js`` can handle that.  The fix for this is actually an optimization in that we didn't need to call ``.empty()`` where we did.  Only in one circumstance did it actually make sense to do so.  So I moved the call to ``.empty()`` there.