Bug 1197796 - Set appropriate max-age for collectstatic generated files
Now that we're using WhiteNoise's `GzipManifestStaticFilesStorage` to
produce hashed (and gzipped) versions of Django static files, we need
to modify `CustomWhiteNoise` so that it supports not only
grunt-cache-busting's style of hash filenames, but those generated by
`GzipManifestStaticFilesStorage` too. For the latter we fall back to the
default DjangoWhiteNoise `is_immutable_file()` method.

The grunt-cache-busting's regex has also been made more strict, to
reduce the chance of false positives, since the hashes are always 32
characters in length.